Refactor API into a Flask Blueprint with a URL prefix
