+++ servicex/api/transformer_file_complete.py
@@ -28,7 +28,7 @@
 from flask import request, current_app
 
 from servicex.models import TransformRequest, TransformationResult, DatasetFile, db
-from servicex.resources.servicex_resource import ServiceXResource
+from servicex.api.servicex_resource import ServiceXResource
 
 
 class TransformerFileComplete(ServiceXResource):
