+++ servicex/api/users/delete_user.py
@@ -3,7 +3,7 @@ import traceback
 import sys
 from flask_jwt_extended import (jwt_optional)
 
-from servicex.resources.servicex_resource import ServiceXResource
+from servicex.api.servicex_resource import ServiceXResource
 from servicex.models import UserModel
