+++ tests/api/test_transform_file_complete.py
@@ -63,7 +63,7 @@ class TestTransformFileComplete(ResourceTestBase):
 
         client = self._test_client(transformation_manager=mock_transformer_manager,
                                    rabbit_adaptor=mock_rabbit_adaptor)
-        response = client.put('/servicex/internal/transformation/1234/file-complete',
+        response = client.put('/api/v1/servicex/internal/transformation/1234/file-complete',
                               json=self._generate_file_complete_request())
         assert response.status_code == 200
         mock_transform_request_read.assert_called_with('1234')
@@ -89,7 +89,7 @@ class TestTransformFileComplete(ResourceTestBase):
 
         client = self._test_client(transformation_manager=mock_transformer_manager,
                                    rabbit_adaptor=mock_rabbit_adaptor)
-        response = client.put('/servicex/internal/transformation/1234/file-complete',
+        response = client.put('/api/v1/servicex/internal/transformation/1234/file-complete',
                               json=self._generate_file_complete_request())
 
         assert response.status_code == 200
@@ -116,7 +116,7 @@ class TestTransformFileComplete(ResourceTestBase):
 
         client = self._test_client(transformation_manager=mock_transformer_manager,
                                    rabbit_adaptor=mock_rabbit_adaptor)
-        response = client.put('/servicex/internal/transformation/1234/file-complete',
+        response = client.put('/api/v1/servicex/internal/transformation/1234/file-complete',
                               json=self._generate_file_complete_request())
 
         assert response.status_code == 200
@@ -166,7 +166,7 @@ class TestTransformFileComplete(ResourceTestBase):
         client = self._test_client(transformation_manager=mock_transformer_manager,
                                    rabbit_adaptor=mock_rabbit_adaptor,
                                    elasticsearch_adapter=mock_elasticsearch_adapter)
-        response = client.put('/servicex/internal/transformation/1234/file-complete',
+        response = client.put('/api/v1/servicex/internal/transformation/1234/file-complete',
                               json=self._generate_file_complete_request())
 
         assert response.status_code == 200
@@ -231,7 +231,7 @@ class TestTransformFileComplete(ResourceTestBase):
         client = self._test_client(transformation_manager=mock_transformer_manager,
                                    rabbit_adaptor=mock_rabbit_adaptor,
                                    elasticsearch_adapter=mock_elasticsearch_adapter)
-        response = client.put('/servicex/internal/transformation/1234/file-complete',
+        response = client.put('/api/v1/servicex/internal/transformation/1234/file-complete',
                               json=self._generate_file_complete_request())
 
         assert response.status_code == 200
