+++ tests/api/test_transformer_start.py
@@ -55,7 +55,7 @@ class TestTransformationStart(ResourceTestBase):
             mock_transformer_manager,
             mock_rabbit_adaptor)
 
-        response = client.post('/servicex/internal/transformation/1234/start',
+        response = client.post('/api/v1/servicex/internal/transformation/1234/start',
                                json={
                                    'info': {
                                        'max-event-size': 4567}
@@ -106,7 +106,7 @@ class TestTransformationStart(ResourceTestBase):
             mock_transformer_manager,
             mock_rabbit_adaptor)
 
-        response = client.post('/servicex/internal/transformation/1234/start',
+        response = client.post('/api/v1/servicex/internal/transformation/1234/start',
                                json={'max-event-size': 4567})
         assert response.status_code == 200
         mock_transform_request_read.assert_called_with('1234')
